requirements.txt
adal==1.1.0
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3.1
Flask==1.0.2
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pycparser==2.19
PyJWT==1.6.4
python-dateutil==2.7.3
python-dotenv==0.9.1
requests==2.19.1
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
